feat: add mutation and hook to save errors
